refactor: 채팅 모달 컴포넌트 분리 #184
